separate headfile operators.hpp into multiple files. Each file for one operator
