+++ src/eliminate_identity.cpp
@@ -1,7 +1,7 @@
 #include <migraphx/eliminate_identity.hpp>
 #include <migraphx/program.hpp>
 #include <migraphx/instruction.hpp>
-#include <migraphx/operators.hpp>
+//#include <migraphx/op/operators.hpp>
 #include <migraphx/iterator_for.hpp>
 #include <migraphx/stringutils.hpp>
 #include <utility>
