+++ src/include/migraphx/op/gru.hpp
@@ -3,6 +3,8 @@
 
 #include <array>
 #include <migraphx/op/common.hpp>
+#include <migraphx/op/tanh.hpp>
+#include <migraphx/op/sigmoid.hpp>
 #include <migraphx/operation.hpp>
 #include <migraphx/check_shapes.hpp>
 #include <migraphx/stringutils.hpp>
