+++ src/include/migraphx/op/unary.hpp
@@ -25,81 +25,6 @@ struct unary
     }
 };
 
-struct abs : unary
-{
-    std::string name() const { return "abs"; }
-};
-
-struct exp : unary
-{
-    std::string name() const { return "exp"; }
-};
-
-struct log : unary
-{
-    std::string name() const { return "log"; }
-};
-
-struct sin : unary
-{
-    std::string name() const { return "sin"; }
-};
-
-struct cos : unary
-{
-    std::string name() const { return "cos"; }
-};
-
-struct tan : unary
-{
-    std::string name() const { return "tan"; }
-};
-
-struct asin : unary
-{
-    std::string name() const { return "asin"; }
-};
-
-struct acos : unary
-{
-    std::string name() const { return "acos"; }
-};
-
-struct atan : unary
-{
-    std::string name() const { return "atan"; }
-};
-
-struct sinh : unary
-{
-    std::string name() const { return "sinh"; }
-};
-
-struct cosh : unary
-{
-    std::string name() const { return "cosh"; }
-};
-
-struct tanh : unary
-{
-    std::string name() const { return "tanh"; }
-};
-
-struct sigmoid : unary
-{
-    std::string name() const { return "sigmoid"; }
-};
-
-struct neg : unary
-{
-    std::string name() const { return "neg"; }
-};
-
-struct relu : unary
-{
-    std::string name() const { return "relu"; }
-};
-
 } // namespace op
 } // namespace MIGRAPHX_INLINE_NS
 } // namespace migraphx
